test(contract): share StubResponse via a types crate, drop the mirror
StubResponse was declared twice (stub contract + test mirror) kept
aligned by a KEEP-IN-SYNC comment and a discriminant-pinning test,
because importing the #[near] stub crate as a dep breaks
cargo test --all-features (duplicate contract-ABI symbol + abi-feature
unification; see PR #3664).

Extract it into a new plain-lib crate test-tee-verifier-types that both
the stub and the contract's test binary depend on. A non-#[near] crate
emits no contract-ABI symbol, so it can be a dev-dep of mpc-contract
without the collision, giving a real single source of truth: the mirror,
the sync comment, and the stub_response_discriminants test are removed.

Verified in nix: cargo test --no-run --all-features -p mpc-contract
links cleanly (the step that regressed in #3664), and the sandbox
tee_verifier suite passes driven through the shared type.
