+++ Cargo.toml
@@ -38,6 +38,7 @@ members = [
     "crates/test-parallel-contract",
     "crates/test-port-allocator",
     "crates/test-tee-verifier",
+    "crates/test-tee-verifier-types",
     "crates/test-utils",
     "crates/threshold-signatures",
     "crates/tls",
@@ -77,6 +78,7 @@ tee-authority = { path = "crates/tee-authority" }
 tee-verifier-conversions = { path = "crates/tee-verifier-conversions" }
 tee-verifier-interface = { path = "crates/tee-verifier-interface" }
 test-port-allocator = { path = "crates/test-port-allocator" }
+test-tee-verifier-types = { path = "crates/test-tee-verifier-types" }
 test-utils = { path = "crates/test-utils" }
 threshold-signatures = { path = "crates/threshold-signatures" }
