+++ crates/contract/Cargo.toml
@@ -140,6 +140,7 @@ rand_core = { workspace = true }
 rstest = { workspace = true }
 sha2 = { workspace = true }
 signature = { workspace = true }
+test-tee-verifier-types = { workspace = true }
 test-utils = { workspace = true }
 threshold-signatures = { workspace = true }
 tokio = { workspace = true }
