+++ crates/test-tee-verifier-types/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "test-tee-verifier-types"
+version = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+# Wire types shared between the `test-tee-verifier` stub contract and the
+# `mpc-contract` sandbox tests that drive it. A plain lib (no `#[near]`) so both
+# a contract crate and a test binary can depend on it without the duplicate-ABI
+# symbol / `--all-features` collision that importing the stub crate itself would
+# cause (see docs / the mpc-contract sandbox tests).
+
+[features]
+# Mirrors the stub's `abi` feature: derives the borsh schema on the wire types so
+# the stub's ABI generation can include them.
+abi = ["borsh/unstable__schema", "tee-verifier-interface/borsh-schema"]
+
+[dependencies]
+borsh = { workspace = true }
+tee-verifier-interface = { workspace = true }
+
+[lints]
+workspace = true
