+++ crates/test-tee-verifier/src/lib.rs
@@ -8,42 +8,16 @@
 //! [`StubResponse::Rejected`] verdict, or a panic (the no-verdict /
 //! verifier-unreachable path).
 
-use borsh::{BorshDeserialize, BorshSerialize};
 use near_sdk::{env, near};
 use tee_verifier_interface::{Collateral, QuoteBytes, VerificationResult, VerifierError};
+use test_tee_verifier_types::StubResponse;
 
-// Match the real verifier's getrandom handling on wasm so the crate links.
 #[cfg(target_arch = "wasm32")]
 fn randomness_unsupported(_buf: &mut [u8]) -> Result<(), getrandom::Error> {
     Err(getrandom::Error::UNSUPPORTED)
 }
 #[cfg(target_arch = "wasm32")]
 getrandom::register_custom_getrandom!(randomness_unsupported);
-
-/// What the stub's [`TestTeeVerifier::verify_quote`] should do, chosen by the
-/// test at deploy time.
-#[expect(clippy::large_enum_variant)]
-// KEEP THE VARIANT ORDER IN SYNC with the `StubResponse` mirror in
-// `crates/contract/tests/sandbox/tee_verifier.rs`: the test serializes with that
-// copy and this contract deserializes with this one, so the Borsh discriminants
-// (declaration index) must match. That mirror's `stub_response_discriminants`
-// test pins the indices.
-#[derive(Debug, Clone, BorshSerialize, BorshDeserialize)]
-#[cfg_attr(
-    all(feature = "abi", not(target_arch = "wasm32")),
-    derive(borsh::BorshSchema)
-)]
-pub enum StubResponse {
-    /// Return [`VerificationResult::Verified`] with this exact report. Tests that
-    /// want the post-DCAP checks to pass supply the report obtained from the
-    /// real fixture quote.
-    Verified(tee_verifier_interface::VerifiedReport),
-    /// Return [`VerificationResult::Rejected`] with this reason.
-    Rejected(String),
-    /// Panic, simulating an unreachable / crashing verifier (the no-verdict
-    /// path that mpc-contract resolves via the yield timeout).
-    Panic,
-}
 
 #[derive(Debug)]
 #[near(contract_state)]
